// components/home/CubeScanner/types.ts
export type Phase =
  | 'cube_appearing'
  | 'rotating'
  | 'gathering_and_aligning'
  | 'waiting'
  | 'zooming_out'
  | 'complete'

export interface CubeScannerConfig {
  SCANBAR_SPEED: number
  SCAN_START_Y: number
  SCAN_END_Y: number
  PARTICLE_COUNT: number
  PARTICLE_PREFADE_MARGIN: number
  PARTICLE_START_Y: number
  PARTICLE_GATHER_MS: number
  PARTICLE_SIZE_START: number
  PARTICLE_VISIBLE_OPACITY: number
  PARTICLE_FADE_AFTER_FLASH_MS: number
  ALIGN_DURATION_MS: number
  WAIT_AFTER_ALIGN_MS: number
  SHAKE_START_THRESHOLD: number
  SHAKE_INTENSITY: number
  SHAKE_FREQUENCY: number
  FLASH_DURATION_MS: number

  CUBE_APPEAR_DURATION: number
  SCAN_HOLD_BEFORE_DISAPPEAR: number
  SCAN_DISAPPEAR_DURATION: number
  ZOOM_OUT_DURATION_MS: number

  CONVERGE_WHOOSH_START: number
  CONVERGE_SNAP_AT: number
}

export const DefaultConfig: CubeScannerConfig = {
  SCANBAR_SPEED: 0.012,
  SCAN_START_Y: 1.0,
  SCAN_END_Y: -1.3333333333,
  PARTICLE_COUNT: 500,
  PARTICLE_PREFADE_MARGIN: 0.25,
  PARTICLE_START_Y: -1.2,
  PARTICLE_GATHER_MS: 2000,
  PARTICLE_SIZE_START: 1,
  PARTICLE_VISIBLE_OPACITY: 0.9,
  PARTICLE_FADE_AFTER_FLASH_MS: 250,
  ALIGN_DURATION_MS: 2000,
  WAIT_AFTER_ALIGN_MS: 1000,
  SHAKE_START_THRESHOLD: 0.65,
  SHAKE_INTENSITY: 0.08,
  SHAKE_FREQUENCY: 25,
  FLASH_DURATION_MS: 200,

  CUBE_APPEAR_DURATION: 800,
  SCAN_HOLD_BEFORE_DISAPPEAR: 500,
  SCAN_DISAPPEAR_DURATION: 800,
  ZOOM_OUT_DURATION_MS: 800,

  CONVERGE_WHOOSH_START: 0.82,
  CONVERGE_SNAP_AT: 0.98,
}